requirements/base.txt
# Base requierment packages

CherryPy
cryptography
Mako
matplotlib
passlib
python-dateutil
aiosqlite
